General Clenup
- Split large lib-file into smaller files
- Renamed getters to follow rust naming convention
- Pretty debug print in examples
- No hardcoded dependency versions in lib
