+++ piped/Cargo.toml
@@ -11,6 +11,6 @@ version = "0.0.0"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-reqwest = "0.11.4"
-serde = {version = "1.0.130", features = ["derive"]}
-serde_json = "1.0.68"
+reqwest = "^0.11"
+serde = {version = "^1.0", features = ["derive"]}
+serde_json = "^1.0"
